split up tests in ms-authenticate
